refactor(cmd): decompose mount.go into focused files [mache-godfile-mount]
Split cmd/mount.go (1026 LOC) into a top-level command file plus 5
single-concern files, with hermetic tests for the directly-flagged
surfaces.

Final LOC: cmd/mount.go 509 (init, rootCmd RunE, versionCmd, Execute) +
- cmd/mount_control.go (178): Control Block hot-swap path
  (mountControl, mountControlWritable, mountWritableNFS, shortRoot)
- cmd/mount_nfs.go (140): NFS server + write-back closure (mountNFS)
- cmd/mount_commands.go (192): list / unmount / clean subcommands
- cmd/mount_inference.go (36): inferFromTreeSitterFile helper
- cmd/mount_extractor.go (35): newCallExtractor (sync.Pool of parsers)

Tests added for the new file surfaces:
- cmd/mount_control_test.go: 3 cases for shortRoot (zero / non-zero /
  single-byte boundary)
- cmd/mount_inference_test.go: 2 cases for inferFromTreeSitterFile
  (real Go source / missing file)
- cmd/mount_extractor_test.go: 2 cases for newCallExtractor (unknown
  language short-circuit / empty Go content)

Defensive nil-tree and parse-error branches that tree-sitter cannot
reach with a valid grammar are marked // coverage:ignore.

Coverage gate (rename-threshold 50): exit 0.

Pure mechanical move — no behavior changes, no public surface
changes, existing tests reference moved functions by name and pass
unchanged.

The rootCmd RunE block (~430 lines: flag setup, schema load/infer,
backend selection, ingest, mount) was left in cmd/mount.go. Splitting
it further requires lifting the closure's captured locals into a
helper struct (not a mechanical move).
