+++ cmd/mount.go
@@ -1,33 +1,23 @@
 package cmd
 
 import (
-	"context"
 	"crypto/sha256"
 	"encoding/json"
 	"fmt"
 	"io"
 	"log"
 	"os"
-	"os/signal"
 	"path/filepath"
-	"strings"
-	"sync"
-	"syscall"
 	"time"
 
 	"github.com/agentic-research/mache/api"
-	"github.com/agentic-research/mache/internal/control"
 	"github.com/agentic-research/mache/internal/graph"
 	"github.com/agentic-research/mache/internal/ingest"
 	"github.com/agentic-research/mache/internal/lang"
 	"github.com/agentic-research/mache/internal/lattice"
 	"github.com/agentic-research/mache/internal/leyline"
-	"github.com/agentic-research/mache/internal/linter"
 	"github.com/agentic-research/mache/internal/materialize"
-	"github.com/agentic-research/mache/internal/nfsmount"
 	machetmpl "github.com/agentic-research/mache/internal/template"
-	"github.com/agentic-research/mache/internal/writeback"
-	sitter "github.com/smacker/go-tree-sitter"
 	"github.com/spf13/cobra"
 )
 
@@ -507,513 +497,6 @@ var rootCmd = &cobra.Command{
 		go leyline.TriggerEmbedding(g, 100)
 
 		return mountNFS(schema, g, engine, mountPoint, writable, promptContent)
-	},
-}
-
-// mountControl starts Mache in hot-swap mode using the Control Block.
-func mountControl(path string, schema *api.Topology, mountPoint string) error {
-	ctrl, err := control.OpenOrCreate(path)
-	if err != nil {
-		return fmt.Errorf("open control: %w", err)
-	}
-	defer func() { _ = ctrl.Close() }()
-
-	// Initial Load: read the controller's current arena pointer + root.
-	// The root is the substrate identity; a zero root means no snapshot
-	// has been published yet (fresh controller).
-	initialRoot := ctrl.GetCurrentRoot()
-	arenaPath := ctrl.GetArenaPath()
-	log.Printf("Control Block: root %s -> %s", shortRoot(initialRoot), arenaPath)
-
-	// Wait for first valid arena if empty
-	if arenaPath == "" {
-		log.Println("Waiting for initial arena...")
-		deadline := time.After(30 * time.Second)
-		for {
-			select {
-			case <-deadline:
-				return fmt.Errorf("timed out waiting for initial arena (30s)")
-			default:
-			}
-			if p := ctrl.GetArenaPath(); p != "" {
-				arenaPath = p
-				break
-			}
-			time.Sleep(100 * time.Millisecond)
-		}
-	}
-
-	// Writable mode still extracts the active buffer to a temp file —
-	// WritableGraph splices edits into a real SQLite connection before
-	// flushing back to the arena, and the daemon doesn't expose write
-	// ops over UDS. Read-only mode skips this entirely and queries the
-	// daemon via UDS.
-	if writable {
-		log.Println("Waiting for valid arena header (writable mode)...")
-		var dbPath string
-		deadline := time.After(30 * time.Second)
-		for {
-			var extractErr error
-			dbPath, extractErr = graph.ExtractActiveDB(arenaPath)
-			if extractErr == nil {
-				break
-			}
-			select {
-			case <-deadline:
-				return fmt.Errorf("timed out waiting for valid arena header (30s): %w", extractErr)
-			default:
-			}
-			time.Sleep(500 * time.Millisecond)
-		}
-		log.Println("Arena header valid. Initializing writable graph.")
-		return mountControlWritable(dbPath, arenaPath, schema, ctrl, mountPoint)
-	}
-
-	// Read-only mode: query the daemon over UDS. The daemon owns SQLite
-	// (zero-copy via sqlite3_deserialize on the arena); mache never opens
-	// a local SQLite file. Hot-swap is implicit — each query reflects
-	// the daemon's current snapshot, so root bumps don't need a graph
-	// swap on the mache side. LLO 0.2.2 added `find_callees` so callees-
-	// dependent features (`/callees` vfs, `find_impact`, `find_smells`
-	// call-graph rules) work over UDS too. Closes mache-98b9bf.
-	sockPath := strings.TrimSuffix(path, ".ctrl") + ".sock"
-	log.Printf("Waiting for daemon socket %s...", sockPath)
-	// Loop on the dial itself rather than stat'ing the socket file:
-	// (a) a stale leftover .sock from a previous daemon will pass stat
-	//     but fail to dial — retry lets us wait through the new daemon
-	//     re-creating it;
-	// (b) the file can exist before the daemon has actually started
-	//     accepting connections (bind-then-listen race), so the first
-	//     dial often hits ECONNREFUSED even when subsequent ones work.
-	// 30s deadline matches the prior stat-based loop.
-	var g *udsGraph
-	sockDeadline := time.Now().Add(30 * time.Second)
-	for {
-		var dialErr error
-		g, dialErr = newUDSGraph(sockPath)
-		if dialErr == nil {
-			break
-		}
-		if time.Now().After(sockDeadline) {
-			return fmt.Errorf("timed out connecting to daemon socket %s after 30s: %w", sockPath, dialErr)
-		}
-		time.Sleep(100 * time.Millisecond)
-	}
-	defer func() { _ = g.Close() }()
-	log.Printf("Connected to ley-line daemon at %s", sockPath)
-
-	return mountNFS(schema, g, nil, mountPoint, false, nil)
-}
-
-// shortRoot renders the first 4 bytes of a 32-byte BLAKE3 root as 8 hex
-// chars, or "<zero>" for the all-zero sentinel — readable enough for log
-// lines without dragging full 64-char hex everywhere.
-func shortRoot(root [32]byte) string {
-	if control.IsZeroRoot(root) {
-		return "<zero>"
-	}
-	return fmt.Sprintf("%02x%02x%02x%02x", root[0], root[1], root[2], root[3])
-}
-
-// mountControlWritable opens the extracted DB in read-write mode and
-// wires a WritableGraph + ArenaFlusher for arena write-back.
-func mountControlWritable(masterDBPath, arenaPath string, schema *api.Topology, ctrl *control.Controller, mountPoint string) error {
-	flusher := graph.NewArenaFlusher(arenaPath, masterDBPath, ctrl)
-	flusher.Start(100 * time.Millisecond)
-	defer func() { _ = flusher.Close() }() // final flush on unmount
-
-	wg, err := graph.OpenWritableGraph(masterDBPath, schema, machetmpl.Render, flusher)
-	if err != nil {
-		return fmt.Errorf("open writable graph: %w", err)
-	}
-	defer func() { _ = wg.Close() }()
-
-	log.Println("Writable arena mode: edits write to master DB and flush to arena (100ms coalesce).")
-
-	return mountWritableNFS(schema, wg, mountPoint)
-}
-
-// mountWritableNFS mounts a WritableGraph via NFS with arena write-back.
-func mountWritableNFS(schema *api.Topology, wg *graph.WritableGraph, mountPoint string) error {
-	graphFs := nfsmount.NewGraphFS(wg, schema)
-
-	graphFs.SetWriteBack(func(nodeID string, origin graph.SourceOrigin, content []byte) error {
-		// Update DB record, then request coalesced arena flush (non-blocking).
-		if err := wg.UpdateRecord(nodeID, content); err != nil {
-			return fmt.Errorf("update record: %w", err)
-		}
-		wg.Flush() // coalesced — actual I/O on next tick
-		return nil
-	})
-
-	srv, err := nfsmount.NewServer(graphFs)
-	if err != nil {
-		return fmt.Errorf("start NFS server: %w", err)
-	}
-	defer func() { _ = srv.Close() }()
-
-	log.Printf("Mounting mache at %s (NFS on localhost:%d)...", mountPoint, srv.Port())
-
-	if err := nfsmount.Mount(srv.Port(), mountPoint, true, nfsOpts); err != nil {
-		return err
-	}
-	log.Print("Mounted (writable). Press Ctrl-C to unmount.")
-
-	sig := make(chan os.Signal, 1)
-	signal.Notify(sig, syscall.SIGINT, syscall.SIGTERM)
-	<-sig
-
-	log.Printf("Unmounting %s...", mountPoint)
-	if err := nfsmount.Unmount(mountPoint); err != nil {
-		log.Printf("Warning: unmount failed: %v", err)
-		log.Printf("Run manually: sudo umount %s", mountPoint)
-	}
-	return nil
-}
-
-// mountNFS starts an NFS server backed by GraphFS and mounts it.
-func mountNFS(schema *api.Topology, g graph.Graph, engine *ingest.Engine, mountPoint string, writable bool, promptContent []byte) error {
-	graphFs := nfsmount.NewGraphFS(g, schema)
-	if len(promptContent) > 0 {
-		graphFs.SetPromptContent(promptContent)
-	}
-
-	// Wire write-back if requested (validate → format → splice → surgical update → invalidate)
-	if writable && engine != nil {
-		store, isMemStore := g.(*graph.MemoryStore)
-		graphFs.SetWriteBack(func(nodeID string, origin graph.SourceOrigin, content []byte) error {
-			// Retrieve node to update DraftData
-			node, err := g.GetNode(nodeID)
-			if err != nil {
-				return fmt.Errorf("node not found: %w", err)
-			}
-
-			// 1. Validate syntax before touching source file
-			if err := writeback.Validate(content, origin.FilePath); err != nil {
-				log.Printf("writeback: validation failed for %s: %v (saving draft)", origin.FilePath, err)
-				// Store diagnostic for _diagnostics/ virtual dir
-				if isMemStore {
-					store.WriteStatus.Store(filepath.Dir(nodeID), err.Error())
-					// Save as Draft
-					draft := make([]byte, len(content))
-					copy(draft, content)
-					node.DraftData = draft
-				}
-				return nil
-			}
-
-			// 2. Format in-process (gofumpt for Go, hclwrite for HCL/Terraform)
-			formatted := writeback.FormatBuffer(content, origin.FilePath)
-
-			// Linting (Warning only)
-			if strings.HasSuffix(origin.FilePath, ".go") {
-				if diags, err := linter.Lint(formatted, "go"); err == nil && len(diags) > 0 {
-					var sb strings.Builder
-					for _, d := range diags {
-						sb.WriteString(d.String() + "\n")
-					}
-					store.WriteStatus.Store(filepath.Dir(nodeID)+"/lint", sb.String())
-				} else {
-					store.WriteStatus.Delete(filepath.Dir(nodeID) + "/lint")
-				}
-			}
-
-			// 3. Splice formatted content into source file
-			oldLen := origin.EndByte - origin.StartByte
-			if err := writeback.Splice(origin, formatted); err != nil {
-				return err
-			}
-
-			// 4. Surgical node update — no re-ingest
-			newOrigin := &graph.SourceOrigin{
-				FilePath:  origin.FilePath,
-				StartByte: origin.StartByte,
-				EndByte:   origin.StartByte + uint32(len(formatted)),
-			}
-			if isMemStore {
-				delta := int32(len(formatted)) - int32(oldLen)
-				if delta != 0 {
-					store.ShiftOrigins(origin.FilePath, origin.EndByte, delta)
-				}
-				// Use source file mtime for deterministic timestamps
-				modTime := time.Now()
-				if fi, err := os.Stat(origin.FilePath); err == nil {
-					modTime = fi.ModTime()
-				}
-				if err := store.UpdateNodeContent(nodeID, formatted, newOrigin, modTime); err != nil {
-					// Splice already touched disk; the graph is now stale.
-					// Surface via WriteStatus so the _diagnostics virtual
-					// dir reflects reality (file on disk is correct,
-					// in-memory graph is wrong, re-read to refresh).
-					log.Printf("writeback: graph update failed after splice for %s: %v", nodeID, err)
-					store.WriteStatus.Store(filepath.Dir(nodeID),
-						fmt.Sprintf("graph stale (splice succeeded, graph update failed): %v", err))
-				} else {
-					store.WriteStatus.Store(filepath.Dir(nodeID), "ok")
-				}
-				store.RecordFileMtime(origin.FilePath, modTime)
-			}
-
-			// 5. Invalidate cached size/content
-			g.Invalidate(nodeID)
-			return nil
-		})
-		log.Println("Write-back enabled: edits will splice into source files.")
-	} else if writable {
-		log.Println("Warning: --writable ignored (only supported for non-.db sources)")
-	}
-
-	srv, err := nfsmount.NewServer(graphFs)
-	if err != nil {
-		return fmt.Errorf("start NFS server: %w", err)
-	}
-	defer func() { _ = srv.Close() }()
-
-	log.Printf("Mounting mache at %s (NFS on localhost:%d)...", mountPoint, srv.Port())
-
-	if err := nfsmount.Mount(srv.Port(), mountPoint, writable, nfsOpts); err != nil {
-		return err
-	}
-	log.Print("Mounted. Press Ctrl-C to unmount.")
-
-	// Block until signal
-	sig := make(chan os.Signal, 1)
-	signal.Notify(sig, syscall.SIGINT, syscall.SIGTERM)
-	<-sig
-
-	log.Printf("Unmounting %s...", mountPoint)
-	if err := nfsmount.Unmount(mountPoint); err != nil {
-		log.Printf("Warning: unmount failed: %v", err)
-		log.Printf("Run manually: sudo umount %s", mountPoint)
-	}
-	return nil
-}
-
-// FUSE backend removed in v0.7.0 (ADR-0006). NFS is the only mount backend.
-// For FUSE mounts, use ley-line-open's `leyline serve`.
-
-// inferFromTreeSitterFile reads a source file, parses it with tree-sitter,
-// and infers a topology schema. Returns an error if parsing fails.
-func inferFromTreeSitterFile(inf *lattice.Inferrer, path string, lang *sitter.Language, label string) (*api.Topology, error) {
-	log.Printf("Inferring schema from %s source via Tree-sitter...", label)
-	start := time.Now()
-	defer func() { log.Printf("Schema inference done in %v", time.Since(start)) }()
-
-	content, err := os.ReadFile(path)
-	if err != nil {
-		return nil, err
-	}
-	parser := sitter.NewParser()
-	parser.SetLanguage(lang)
-	tree, parseErr := parser.ParseCtx(context.Background(), nil, content)
-	if parseErr != nil {
-		return nil, fmt.Errorf("tree-sitter parse failed for %s: %w", path, parseErr)
-	}
-	if tree == nil {
-		return nil, fmt.Errorf("tree-sitter returned nil tree for %s", path)
-	}
-	return inf.InferFromTreeSitter(tree.RootNode())
-}
-
-// newCallExtractor creates a CallExtractor that uses a sync.Pool of tree-sitter
-// parsers to reduce allocation overhead. Safe for concurrent use.
-func newCallExtractor() graph.CallExtractor {
-	walker := ingest.NewSitterWalker()
-	pool := &sync.Pool{
-		New: func() any { return sitter.NewParser() },
-	}
-	return func(content []byte, path, langName string) ([]graph.QualifiedCall, error) {
-		l := lang.ForName(langName)
-		if l == nil {
-			return nil, nil
-		}
-		grammar := l.Grammar()
-		parser := pool.Get().(*sitter.Parser)
-		defer pool.Put(parser)
-		parser.SetLanguage(grammar)
-		tree, _ := parser.ParseCtx(context.Background(), nil, content)
-		if tree == nil {
-			return nil, nil
-		}
-		return walker.ExtractQualifiedCalls(tree.RootNode(), content, grammar, langName)
-	}
-}
-
-var listCmd = &cobra.Command{
-	Use:   "list",
-	Short: "List active mache instances (mounts and MCP servers)",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		mounts, err := listActiveMounts()
-		if err != nil {
-			return err
-		}
-
-		if len(mounts) == 0 {
-			fmt.Println("No active mache instances found.")
-			return nil
-		}
-
-		fmt.Printf("%-20s %-12s %-10s %-40s %s\n", "NAME", "TYPE", "PID", "SOURCE", "STATUS")
-		fmt.Println(strings.Repeat("-", 100))
-
-		for _, meta := range mounts {
-			name := filepath.Base(meta.MountPoint)
-			status := "running"
-			if !isProcessRunning(meta.PID) {
-				status = "stale"
-			}
-			typ := meta.Type
-			if typ == "" {
-				typ = "mount" // backwards compat for old sidecars
-			}
-			source := meta.Source
-			if meta.Addr != "" {
-				source = meta.Addr + " " + source
-			}
-			fmt.Printf("%-20s %-12s %-10d %-40s %s\n", name, typ, meta.PID, source, status)
-		}
-
-		return nil
-	},
-}
-
-var unmountCmd = &cobra.Command{
-	Use:   "unmount <mount-name>",
-	Short: "Unmount and stop a mache mount",
-	Args:  cobra.ExactArgs(1),
-	RunE: func(cmd *cobra.Command, args []string) error {
-		mountName := args[0]
-
-		// Resolve mount point (support both short name and full path)
-		var mountPoint string
-		if filepath.IsAbs(mountName) {
-			mountPoint = mountName
-		} else {
-			mountsDir, err := getAgentMountsDir()
-			if err != nil {
-				return err
-			}
-			mountPoint = filepath.Join(mountsDir, mountName)
-		}
-
-		// Load metadata. The sidecar exists for agent-mode and serve-
-		// mode mounts; direct `mache <path>` and ley-line --control
-		// mounts have no sidecar. Don't fail when it's missing —
-		// unmount the kernel-side NFS first regardless, then clean up
-		// what we can (mache-fsi: 'No agent metadata' case).
-		meta, metaErr := loadMountMetadata(mountPoint)
-
-		isMCPServe := meta != nil && (meta.Type == "mcp-http" || meta.Type == "mcp-stdio")
-
-		// Step 1: kernel-side unmount FIRST (for NFS mounts only).
-		// SIGTERM-then-RemoveAll without an explicit umount races —
-		// if the daemon doesn't finish its own unmount before
-		// RemoveAll runs, we leave a stuck NFS mount that only
-		// `umount -f` can clear (mache-fsi: 'Race' + 'Orphaned mount'
-		// cases). Unmount is idempotent — repeating on an already-
-		// unmounted point logs a warning and we continue.
-		if !isMCPServe {
-			if err := nfsmount.Unmount(mountPoint); err != nil {
-				log.Printf("Note: nfsmount.Unmount(%s): %v (continuing)", mountPoint, err)
-			}
-		}
-
-		// Step 2: SIGTERM the owning process (only when we have its
-		// PID via the sidecar). For control / sidecar-less mounts
-		// the user is expected to stop the daemon separately.
-		if meta != nil && isProcessRunning(meta.PID) {
-			process, err := os.FindProcess(meta.PID)
-			if err == nil {
-				log.Printf("Stopping mache process (PID %d)...", meta.PID)
-				_ = process.Signal(syscall.SIGTERM)
-
-				// Wait briefly for graceful shutdown
-				time.Sleep(2 * time.Second)
-
-				if isProcessRunning(meta.PID) {
-					log.Printf("Process still running, sending SIGKILL...")
-					_ = process.Signal(syscall.SIGKILL)
-				}
-			} else {
-				log.Printf("Note: failed to find process %d: %v", meta.PID, err)
-			}
-		}
-
-		// Step 3: clean up mount directory and sidecar.
-		log.Printf("Removing mount directory: %s", mountPoint)
-		if err := os.RemoveAll(mountPoint); err != nil {
-			return fmt.Errorf("failed to remove mount directory: %w", err)
-		}
-		_ = os.Remove(sidecarPath(mountPoint))
-
-		if metaErr != nil {
-			log.Printf("Mount stopped (no sidecar metadata; resolved by direct unmount).")
-		} else {
-			log.Println("Mount stopped successfully.")
-		}
-		return nil
-	},
-}
-
-var cleanCmd = &cobra.Command{
-	Use:   "clean",
-	Short: "Remove stale mache mounts and orphaned snapshots",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		mounts, err := listActiveMounts()
-		if err != nil {
-			return err
-		}
-
-		cleaned := 0
-		for _, meta := range mounts {
-			if !isProcessRunning(meta.PID) {
-				log.Printf("Removing stale mount: %s (PID %d was not running)",
-					filepath.Base(meta.MountPoint), meta.PID)
-				if err := os.RemoveAll(meta.MountPoint); err != nil {
-					log.Printf("Warning: failed to remove %s: %v", meta.MountPoint, err)
-				} else {
-					_ = os.Remove(sidecarPath(meta.MountPoint))
-					cleaned++
-				}
-			}
-		}
-
-		// Clean orphaned snapshots (snap-<PID>-* where PID is dead)
-		snapDir := filepath.Join(os.TempDir(), "mache", "snapshots")
-		if entries, err := os.ReadDir(snapDir); err == nil {
-			for _, entry := range entries {
-				name := entry.Name()
-				if !strings.HasPrefix(name, "snap-") {
-					continue
-				}
-				// Parse PID from snap-<PID>-<name>
-				parts := strings.SplitN(strings.TrimPrefix(name, "snap-"), "-", 2)
-				if len(parts) < 2 {
-					continue
-				}
-				var pid int
-				if _, err := fmt.Sscanf(parts[0], "%d", &pid); err != nil {
-					continue
-				}
-				if !isProcessRunning(pid) {
-					snapPath := filepath.Join(snapDir, name)
-					log.Printf("Removing orphaned snapshot: %s (PID %d was not running)", name, pid)
-					if err := os.RemoveAll(snapPath); err != nil {
-						log.Printf("Warning: failed to remove %s: %v", snapPath, err)
-					} else {
-						cleaned++
-					}
-				}
-			}
-		}
-
-		if cleaned == 0 {
-			log.Println("No stale mounts or orphaned snapshots found.")
-		} else {
-			log.Printf("Cleaned %d stale item(s).", cleaned)
-		}
-
-		return nil
 	},
 }
